Reverting type definitions for `Pivot` config.
