+++ graylog2-web-interface/src/views/logic/aggregationbuilder/Pivot.ts
@@ -15,22 +15,7 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 
-export type DatePivotConfig = {
-  interval: {
-    type: 'auto',
-    scaling: number
-  } | {
-    type: 'timeunit',
-    value: number,
-    unit: string
-  }
-}
-
-export type ValuesPivotConfig = {
-  limit: number
-}
-
-type ConfigType = DatePivotConfig | ValuesPivotConfig;
+type ConfigType = { [key: string]: unknown };
 
 export const DateType = 'time';
 export const ValuesType = 'values';
